# Greatly optimize (geohashes-intersecting) for large polygons. #25