* US polygon to L3 geohash went from 350s to 19s (1842%) speedup
* CA polygon to L3 geohash went from 3s to 430ms (698%) speedup