@worace 2018-07-10
oh dang is it repeating the shape conversion over and over?